Implement year-progress-bar-section
It's important to reguarly get a glance of how much more time is left in the year.
This helps set realistic goals and focus on what matters.

We should consider (future feature): Programatically being deliberate about when the user see this bar.
For instance, we could hide it if we've recently shown it to the user.